support replay of incomplete exposures

+++ desietc/offline.py
@@ -17,7 +17,7 @@ import desietc.sky
 import desietc.plot
 
 def replay_exposure(ETC, path, expid, outpath, teff=1000, ttype='DARK', cutoff=10000, cosmic=500,
-                    maxsplit=3, splittable=True, overwrite=False, dry_run=False):
+                    maxsplit=3, splittable=True, overwrite=False, dry_run=False, only_complete=True):
     """Recreate the online ETC processing of an exposure by replaying the
     FITS files stored to disk.
     """
@@ -28,10 +28,15 @@ def replay_exposure(ETC, path, expid, outpath, teff=1000, ttype='DARK', cutoff=1
         logging.error(f'No exposure directory found.')
         return False
     # What data is available for this exposure?
+    acq_path = exppath / f'guide-{exptag}-0000.fits.fz'
     gfa_path = exppath / f'guide-{exptag}.fits.fz'
     sky_path = exppath / f'sky-{exptag}.fits.fz'
     desi_path = exppath / f'desi-{exptag}.fits.fz'
+    fassign_paths = list(exppath.glob('fiberassign-??????.fits.gz'))
     missing = 0
+    if not acq_path.exists():
+        logging.warn(f'Missing acquisition image: {acq_path}.')
+        missing += 1
     if not gfa_path.exists():
         logging.warn(f'Missing GFA data cube: {gfa_path}.')
         missing += 1
@@ -41,26 +46,28 @@ def replay_exposure(ETC, path, expid, outpath, teff=1000, ttype='DARK', cutoff=1
     if not desi_path.exists():
         logging.warn(f'Missing DESI exposure: {desi_path}.')
         missing += 1
-    if missing > 0:
+    if len(fassign_paths) != 0:
+        logging.warn(f'Missing fiberassign file.')
+        fassign_path = exppath / '_does_not_exist_'
+        missing += 1
+    else:
+        fassign_path = fassign_paths[1]
+    if only_complete and missing > 0:
         return False
-    # Get the spectrograph info for this exposure.
-    desi_hdr = fitsio.read_header(str(desi_path), ext='SPEC')
-    missing = 0
-    for key in 'NIGHT', 'MJD-OBS', 'TILEID', 'EXPTIME':
-        if key not in desi_hdr:
-            logging.error('DESI exposure missing {key}.')
-            missing += 1
-    if missing > 0:
-        return False
-    night = desi_hdr['NIGHT']
-    desi_mjd_obs = desi_hdr['MJD-OBS']
-    desi_exptime = desi_hdr['EXPTIME']
-    desi_tileid = desi_hdr['TILEID']
-    # Locate the fiberassign file for this tile.
-    fassign_path = exppath / f'fiberassign-{desi_tileid:06d}.fits'
-    if not fassign_path.exists():
-        logging.error(f'Missing fiberassign file: {fassign_path}.')
-        return False
+    if desi_path.exists():
+        # Get the spectrograph info for this exposure.
+        desi_hdr = fitsio.read_header(str(desi_path), ext='SPEC')
+        missing = 0
+        for key in 'NIGHT', 'MJD-OBS', 'TILEID', 'EXPTIME':
+            if key not in desi_hdr:
+                logging.error('DESI exposure missing {key}.')
+                missing += 1
+        if missing > 0:
+            return False
+        night = desi_hdr['NIGHT']
+        desi_mjd_obs = desi_hdr['MJD-OBS']
+        desi_exptime = desi_hdr['EXPTIME']
+        desi_tileid = desi_hdr['TILEID']
     # If this is a dry run, stop now.
     if dry_run:
         return True
@@ -69,38 +76,47 @@ def replay_exposure(ETC, path, expid, outpath, teff=1000, ttype='DARK', cutoff=1
     if not overwrite and exppath_out.exists():
         logging.info(f'Will not overwrite ETC outputs for {expid}.')
         return False
+    # Create the output exposure directory if necessary.
+    exppath_out.mkdir(parents=False, exist_ok=True)
     # Save images with the per-exposure outputs.
     ETC.set_image_path(exppath_out)
     # Get the SKY exposure info for the first available camera.
-    sky_info = None
-    with fitsio.FITS(str(sky_path)) as hdus:
-        for camera in desietc.sky.SkyCamera.sky_names:
-            if camera+'T' not in hdus:
-                continue
-            sky_info = hdus[camera + 'T'].read()
-            break
-    if sky_info is None:
-        logging.error(f'Unable to find SKY exposure info.')
-        return False
+    sky_info = []
+    if sky_path.exists():
+        with fitsio.FITS(str(sky_path)) as hdus:
+            for camera in desietc.sky.SkyCamera.sky_names:
+                if camera+'T' not in hdus:
+                    continue
+                sky_info = hdus[camera + 'T'].read()
+                break
     num_sky_frames = len(sky_info)
     logging.info(f'Exposure has {num_sky_frames} SKY frames.')
     # Get the PM guide stars and GFA exposure info for the first available camera.
-    gfa_info = None
-    with fitsio.FITS(str(gfa_path)) as hdus:
-        # Lookup the PlateMaker guide stars.
-        if 'PMGSTARS' not in hdus:
-            logging.error(f'Missing PMGSTARS HDU: no guide stars specified.')
-            return False
-        pm_info = hdus['PMGSTARS'].read()
-        logging.info(f'Exposure has {len(pm_info[0])} guide stars.')
-        for camera in desietc.gfa.GFACamera.guide_names:
-            if camera+'T' not in hdus:
-                continue
-            gfa_info = hdus[camera + 'T'].read()
-            break
-    if gfa_info is None:
-        logging.error(f'Unable to find GFA exposure info.')
-        return False
+    gfa_info, pm_info = [], None
+    if gfa_path.exists():
+        with fitsio.FITS(str(gfa_path)) as hdus:
+            # Lookup the PlateMaker guide stars.
+            if 'PMGSTARS' not in hdus:
+                logging.error(f'Missing PMGSTARS HDU: no guide stars specified.')
+                return False
+            pm_info = hdus['PMGSTARS'].read()
+            logging.info(f'Exposure has {len(pm_info[0])} guide stars.')
+            for camera in desietc.gfa.GFACamera.guide_names:
+                if camera+'T' not in hdus:
+                    continue
+                gfa_info = hdus[camera + 'T'].read()
+                break
+    elif acq_path.exists():
+        # We have an acq image but no subsequent guider frames.
+        with fitsio.FITS(str(acq_path)) as hdus:
+            for camera in desietc.gfa.GFACamera.guide_names:
+                if camera not in hdus:
+                    continue
+                hdr = hdus[camera].read_header()
+                mjd_obs, exptime = hdr['MJD-OBS'], hdr['EXPTIME']
+                gfa_info = np.array(
+                    [(mjd_obs, exptime)], dtype=[('MJD-OBS', float), ('EXPTIME', float)])
+                break
     num_gfa_frames = len(gfa_info)
     logging.info(f'Exposure has {num_gfa_frames} GFA frames.')
     # Determine the order in which the combined GFA+SKY frames should be fed to the ETC.
@@ -110,54 +126,72 @@ def replay_exposure(ETC, path, expid, outpath, teff=1000, ttype='DARK', cutoff=1
         [ dict(typ='sky', num=n, when=sky_info[n]['MJD-OBS']+sky_info[n]['EXPTIME']/ETC.SECS_PER_DAY)
           for n in range(num_sky_frames) ])
     frames = sorted(frames, key=lambda frame: frame['when'])
-    mjd1 = desi_mjd_obs
-    mjd2 = mjd1 + desi_exptime / ETC.SECS_PER_DAY
+    if len(frames) == 0:
+        logging.error(f'No GFA or SKY frames found.')
+        return False
     # Start the exposure processing.
     timestamp = desietc.util.mjd_to_date(gfa_info[0]['MJD-OBS'] - 1 / ETC.SECS_PER_DAY, utc_offset=0)
     ETC.start_exposure(timestamp, expid, teff, ttype, cutoff, cosmic, maxsplit, splittable)
     # Loop over frames to replay.
     for frame in frames:
-        if frame['when'] > mjd2:
-            break
         if frame['typ'] == 'gfa':
-            data = fits_to_online(gfa_path, desietc.gfa.GFACamera.guide_names, frame['num'])
+            if frame['num'] == 0 and acq_path.exists():
+                data = acq_to_online(acq_path, desietc.gfa.GFACamera.guide_names)
+            else:
+                data = fits_to_online(gfa_path, desietc.gfa.GFACamera.guide_names, frame['num'])
             if frame['num'] == 0:
                 # Process the acquisition image.
                 ETC.process_acquisition(data)
-                # Read the fiber assignments for this tile.
-                ETC.read_fiberassign(fassign_path)
-                # Specify the guide stars.
-                ETC.set_guide_stars(pm_info)
-                # Signal the shutter opening.
-                timestamp = desietc.util.mjd_to_date(desi_mjd_obs, utc_offset=0)
-                ETC.open_shutter(timestamp)
+                if fassign_path.exists():
+                    # Read the fiber assignments for this tile.
+                    ETC.read_fiberassign(fassign_path)
+                if pm_info is not None:
+                    # Specify the guide stars.
+                    ETC.set_guide_stars(pm_info)
+                if desi_path.exists():
+                    # Signal the shutter opening.
+                    timestamp = desietc.util.mjd_to_date(desi_mjd_obs, utc_offset=0)
+                    ETC.open_shutter(timestamp)
             else:
                 # Process the next guide frame.
                 ETC.process_guide_frame(data)
         else: # SKY
             data = fits_to_online(sky_path, ETC.SKY.sky_names, frame['num'])
             ETC.process_sky_frame(data)
-    # Signal the shutter closing.
-    timestamp = desietc.util.mjd_to_date(desi_mjd_obs + desi_exptime / ETC.SECS_PER_DAY, utc_offset=0)
-    ETC.close_shutter(timestamp)
-    # End the exposure.
-    ETC.stop_exposure(timestamp)
-    # Create the output path if necessary.
-    exppath_out.mkdir(parents=False, exist_ok=True)
-    # Save the ETC outputs for this exposure.
-    ETC.save_exposure(exppath_out)
-
-    # Plot the signal and background measurement buffers spanning this exposure.
-    fig, ax = plt.subplots(2, 1, figsize=(9, 9))
-    fig.suptitle(f'ETC Analysis for {ETC.night}/{ETC.exptag}')
-    desietc.plot.plot_measurements(
-        ETC.sky_measurements, mjd1, mjd2, label='SKYCAM Level', ax=ax[0])
-    desietc.plot.plot_measurements(
-        ETC.thru_measurements, mjd1, mjd2, label='GFA Throughput', ax=ax[1])
-    plt.savefig(exppath_out / f'etc-measure-{ETC.exptag}.png')
-    plt.close(fig)
+    if desi_path.exists():
+        # Signal the shutter closing.
+        timestamp = desietc.util.mjd_to_date(desi_mjd_obs + desi_exptime / ETC.SECS_PER_DAY, utc_offset=0)
+        ETC.close_shutter(timestamp)
+        # End the exposure.
+        ETC.stop_exposure(timestamp)
+        # Save the ETC outputs for this exposure.
+        ETC.save_exposure(exppath_out)
+        # Plot the signal and background measurement buffers spanning this exposure.
+        mjd1 = desi_mjd_obs
+        mjd2 = mjd1 + desi_exptime / ETC.SECS_PER_DAY
+        fig, ax = plt.subplots(2, 1, figsize=(9, 9))
+        fig.suptitle(f'ETC Analysis for {ETC.night}/{ETC.exptag}')
+        desietc.plot.plot_measurements(
+            ETC.sky_measurements, mjd1, mjd2, label='SKYCAM Level', ax=ax[0])
+        desietc.plot.plot_measurements(
+            ETC.thru_measurements, mjd1, mjd2, label='GFA Throughput', ax=ax[1])
+        plt.savefig(exppath_out / f'etc-measure-{ETC.exptag}.png')
+        plt.close(fig)
 
     return True
+
+
+def acq_to_online(path, names):
+    online = {}
+    with fitsio.FITS(str(path)) as hdus:
+        online['header']= dict(hdus[0].read_header())
+        for ext in names:
+            if ext not in hdus:
+                continue
+            data = hdus[ext][:,:]
+            hdr = hdus[ext].read_header()
+            online[ext] = dict(header=hdr, data=data)
+    return online
 
 
 def fits_to_online(path, names, frame):
@@ -170,7 +204,18 @@ def fits_to_online(path, names, frame):
         for ext in names:
             if ext not in hdus or ext + 'T' not in hdus:
                 continue
-            data = hdus[ext][frame,:,:][0]
+            dims = hdus[ext].get_dims()
+            print('dims', dims, ('='*40))
+            if len(dims) == 2:
+                if frame != 0:
+                    raise ValueError(f'Requested frame {frame} when no frames present.')
+                data = hdus[ext][:,:]
+            elif len(dims) == 3:
+                if frame >= dims[-1]:
+                    raise ValueError(f'Requested non-existent frame {frame}.')
+                data = hdus[ext][frame,:,:][0]
+            else:
+                raise ValueError(f'Data has invalid dimensions: {dims}.')
             table = hdus[ext + 'T'][frame]
             hdr = {key: table[key] for key in table.dtype.names}
             online[ext] = dict(header=hdr, data=data)
